Seperate RGB Matrix code from RGB Light code in userspace
