+++ users/drashna/drashna.c
@@ -93,8 +93,11 @@ void matrix_init_user(void) {
 __attribute__((weak)) void keyboard_post_init_keymap(void) {}
 
 void keyboard_post_init_user(void) {
-#if defined(RGBLIGHT_ENABLE) || defined(RGB_MATRIX_ENABLE)
-    keyboard_post_init_rgb();
+#if defined(RGBLIGHT_ENABLE)
+    keyboard_post_init_rgb_light();
+#endif
+#if defined(RGB_MATRIX_ENABLE)
+    keyboard_post_init_rgb_matrix();
 #endif
     keyboard_post_init_keymap();
 }
@@ -144,9 +147,12 @@ void matrix_scan_user(void) {
     run_diablo_macro_check();
 #endif  // TAP_DANCE_ENABLE
 
-#if defined(RGBLIGHT_ENABLE) || defined(RGB_MATRIX_ENABLE)
-    matrix_scan_rgb();
+#if defined(RGBLIGHT_ENABLE)
+    matrix_scan_rgb_light();
 #endif  // RGBLIGHT_ENABLE
+#if defined(RGB_MATRIX_ENABLE)
+    matrix_scan_rgb_matrix();
+#endif
 
     matrix_scan_keymap();
 }
@@ -157,8 +163,8 @@ __attribute__((weak)) layer_state_t layer_state_set_keymap(layer_state_t state) 
 // Then runs keymap's layer change check
 layer_state_t layer_state_set_user(layer_state_t state) {
     state = update_tri_layer_state(state, _RAISE, _LOWER, _ADJUST);
-#if defined(RGBLIGHT_ENABLE) || defined(RGB_MATRIX_ENABLE)
-    state = layer_state_set_rgb(state);
+#if defined(RGBLIGHT_ENABLE)
+    state = layer_state_set_rgb_light(state);
 #endif  // RGBLIGHT_ENABLE
     return layer_state_set_keymap(state);
 }
